Extract journal-automation endpoints to keep LedgerEndpoints under size cap
Adding the period-close intake endpoint pushed LedgerEndpoints.cs past
its file-size ratchet baseline (3754 > 3712). Rather than grow the
baseline, move the three journal-automation intake endpoints (dividend,
fee-accrual, period-close) into a LedgerEndpoints.JournalAutomation.cs
partial, matching the WorkstationEndpoints.*.cs pattern. Behavior is
unchanged; the class is now partial and MapLedgerEndpoints calls
MapJournalAutomationEndpoints.
