+++ src/Meridian.Ui.Shared/Endpoints/LedgerEndpoints.cs
@@ -18,7 +18,7 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace Meridian.Ui.Shared.Endpoints;
 
-public static class LedgerEndpoints
+public static partial class LedgerEndpoints
 {
     public static void MapLedgerEndpoints(this WebApplication app, JsonSerializerOptions jsonOptions)
     {
@@ -2015,131 +2015,7 @@ public static class LedgerEndpoints
         .RequireFundScopedWriteTenant()
         .RequireRateLimiting(UiEndpoints.MutationRateLimitPolicy);
 
-        app.MapPost(UiApiRoutes.LedgerJournalAutomationDividendIntake, async (RunDividendDraftIntakeRequest request, HttpContext context) =>
-        {
-            if (!HasLedgerMutationPermission(context))
-            {
-                return EndpointHelpers.Forbidden();
-            }
-
-            var runner = context.RequestServices.GetService<AutomatedJournalIntakeRunner>();
-            if (runner is null)
-            {
-                return ServiceUnavailable();
-            }
-
-            try
-            {
-                var tenantContext = HttpContextWorkstationTenantContextAccessor.Resolve(context);
-                var result = await runner.RunDividendIntakeAsync(request with
-                {
-                    Actor = ResolveMutationActor(context, request.Actor),
-                    TenantId = tenantContext.TenantId,
-                    CompanyId = tenantContext.CompanyId
-                }, context.RequestAborted).ConfigureAwait(false);
-                return Results.Json(result, jsonOptions);
-            }
-            catch (ArgumentException ex)
-            {
-                return Results.BadRequest(new { error = ex.Message });
-            }
-            catch (InvalidOperationException ex)
-            {
-                return Results.Conflict(new { error = ex.Message });
-            }
-        })
-        .WithName("RunLedgerJournalAutomationDividendIntake")
-        .Produces<AutomatedJournalIntakeRunResult>(StatusCodes.Status200OK)
-        .Produces(StatusCodes.Status400BadRequest)
-        .Produces(StatusCodes.Status403Forbidden)
-        .Produces(StatusCodes.Status409Conflict)
-        .Produces(StatusCodes.Status501NotImplemented)
-        .RequireFundScopedWriteTenant()
-        .RequireRateLimiting(UiEndpoints.MutationRateLimitPolicy);
-
-        app.MapPost(UiApiRoutes.LedgerJournalAutomationFeeAccrualIntake, async (RunFeeAccrualDraftIntakeRequest request, HttpContext context) =>
-        {
-            if (!HasLedgerMutationPermission(context))
-            {
-                return EndpointHelpers.Forbidden();
-            }
-
-            var runner = context.RequestServices.GetService<AutomatedJournalIntakeRunner>();
-            if (runner is null)
-            {
-                return ServiceUnavailable();
-            }
-
-            try
-            {
-                var tenantContext = HttpContextWorkstationTenantContextAccessor.Resolve(context);
-                var result = await runner.RunFeeAccrualIntakeAsync(request with
-                {
-                    Actor = ResolveMutationActor(context, request.Actor),
-                    TenantId = tenantContext.TenantId,
-                    CompanyId = tenantContext.CompanyId
-                }, context.RequestAborted).ConfigureAwait(false);
-                return Results.Json(result, jsonOptions);
-            }
-            catch (ArgumentException ex)
-            {
-                return Results.BadRequest(new { error = ex.Message });
-            }
-            catch (InvalidOperationException ex)
-            {
-                return Results.Conflict(new { error = ex.Message });
-            }
-        })
-        .WithName("RunLedgerJournalAutomationFeeAccrualIntake")
-        .Produces<AutomatedJournalIntakeRunResult>(StatusCodes.Status200OK)
-        .Produces(StatusCodes.Status400BadRequest)
-        .Produces(StatusCodes.Status403Forbidden)
-        .Produces(StatusCodes.Status409Conflict)
-        .Produces(StatusCodes.Status501NotImplemented)
-        .RequireFundScopedWriteTenant()
-        .RequireRateLimiting(UiEndpoints.MutationRateLimitPolicy);
-
-        app.MapPost(UiApiRoutes.LedgerJournalAutomationPeriodCloseIntake, async (RunPeriodCloseDraftIntakeRequest request, HttpContext context) =>
-        {
-            if (!HasLedgerMutationPermission(context))
-            {
-                return EndpointHelpers.Forbidden();
-            }
-
-            var runner = context.RequestServices.GetService<AutomatedJournalIntakeRunner>();
-            if (runner is null)
-            {
-                return ServiceUnavailable();
-            }
-
-            try
-            {
-                var tenantContext = HttpContextWorkstationTenantContextAccessor.Resolve(context);
-                var result = await runner.RunPeriodCloseIntakeAsync(request with
-                {
-                    Actor = ResolveMutationActor(context, request.Actor),
-                    TenantId = tenantContext.TenantId,
-                    CompanyId = tenantContext.CompanyId
-                }, context.RequestAborted).ConfigureAwait(false);
-                return Results.Json(result, jsonOptions);
-            }
-            catch (ArgumentException ex)
-            {
-                return Results.BadRequest(new { error = ex.Message });
-            }
-            catch (InvalidOperationException ex)
-            {
-                return Results.Conflict(new { error = ex.Message });
-            }
-        })
-        .WithName("RunLedgerJournalAutomationPeriodCloseIntake")
-        .Produces<AutomatedJournalIntakeRunResult>(StatusCodes.Status200OK)
-        .Produces(StatusCodes.Status400BadRequest)
-        .Produces(StatusCodes.Status403Forbidden)
-        .Produces(StatusCodes.Status409Conflict)
-        .Produces(StatusCodes.Status501NotImplemented)
-        .RequireFundScopedWriteTenant()
-        .RequireRateLimiting(UiEndpoints.MutationRateLimitPolicy);
+        MapJournalAutomationEndpoints(app, jsonOptions);
 
         app.MapPost(UiApiRoutes.LedgerManualJournalEntryEvidence, async (AttachManualJournalEntryEvidenceRequest request, HttpContext context) =>
         {
